move twitter agent types to types.ts
